More rearranging and cleanup of full trainer
